* commit '3e63a427aac965d61a6c7ff073ce21cdbf22d9bf':
  add brackets suggested by Thomas
  prevent casts of RelationVetor<A> to RelationVector<B>
  Add remove() and setWeight() to RelationVector

+++ framework/datastore/include/RelationEntry.h
@@ -26,10 +26,10 @@ namespace Belle2 {
      *  @param obj  The object that is pointed to by the relation.
      *  @param w    The weight of the relation.
      */
-    RelationEntry(TObject* obj = nullptr, double w = 1.0): object(obj), weight(w) {}
+    RelationEntry(TObject* obj = nullptr, float w = 1.0): object(obj), weight(w) {}
 
     TObject* object;  /**< Pointer to the object. */
-    double   weight;  /**< Weight of the relation. */
+    float   weight;  /**< Weight of the relation. */
   };
 }
